Mike | refactoring structure of some of the code, removing unnecessary style tags and using a more natural flexbox layout vs. set %
