+++ src/components/Player/index.js
@@ -1,2 +1,2 @@
 export { default } from './Player';
-export { default as AppBadge } from './AppBadge';
+export { default as AppBadge } from './Header/AppBadge';
